Another Interpolation Module
This module uses the same pipeline as the previous. The only difference is that calculating the optical flow and the remapping are not done using built in functions from OpenCV. They are done from scratch to help anyone who wants to learn how these functions are implemented. This module produces lower quality than the previous module; as the optical flow and the remapping could be optimized to produce better results.